refactor: extract file list case handlers onto fileListState
