+++ internal/menu/executor_files_list.go
@@ -5,21 +5,16 @@ import (
 	"fmt"
 	"io"
 	"log/slog"
-	"os"
-	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/gliderlabs/ssh"
-	"github.com/google/uuid"
 	"golang.org/x/term"
 
 	"github.com/ViSiON-3/vision-3-bbs/internal/ansi"
 	"github.com/ViSiON-3/vision-3-bbs/internal/file"
 	"github.com/ViSiON-3/vision-3-bbs/internal/terminalio"
 	"github.com/ViSiON-3/vision-3-bbs/internal/user"
-	"github.com/ViSiON-3/vision-3-bbs/internal/ziplab"
 )
 
 // fileColumnEnabled returns whether a column should be shown in the classic file listing.
@@ -298,10 +293,7 @@ func runListFiles(c *cmdCtx, args string) (*user.User, string, error) {
 			if st.currentPage < st.totalPages {
 				st.currentPage++
 				// Fetch files for the new page
-				st.filesOnPage, err = st.e.FileMgr.GetFilesForAreaPaginated(st.currentAreaID, st.currentPage, st.filesPerPage)
-				if err != nil {
-					// Log error and potentially return or break the loop
-					slog.Error("failed to get files for page", "node", st.nodeNumber, "page", st.currentPage, "error", err)
+				if fetchErr := st.fetchPage(); fetchErr != nil {
 					// Display error message to user?
 					time.Sleep(1 * time.Second)
 				}
@@ -315,10 +307,7 @@ func runListFiles(c *cmdCtx, args string) (*user.User, string, error) {
 			if st.currentPage > 1 {
 				st.currentPage--
 				// Fetch files for the new page
-				st.filesOnPage, err = st.e.FileMgr.GetFilesForAreaPaginated(st.currentAreaID, st.currentPage, st.filesPerPage)
-				if err != nil {
-					// Log error and potentially return or break the loop
-					slog.Error("failed to get files for page", "node", st.nodeNumber, "page", st.currentPage, "error", err)
+				if fetchErr := st.fetchPage(); fetchErr != nil {
 					// Display error message to user?
 					time.Sleep(1 * time.Second)
 				}
@@ -332,187 +321,21 @@ func runListFiles(c *cmdCtx, args string) (*user.User, string, error) {
 			slog.Debug("user quit LISTFILES", "node", st.nodeNumber)
 			return nil, "", nil // Return to FILEM menu
 		case "D": // Download marked files
-			slog.Debug("user initiated download command", "node", st.nodeNumber, "handle", st.currentUser.Handle, "area", st.currentAreaID)
-
-			// 1. Check if any files are marked
-			if len(st.currentUser.TaggedFileIDs) == 0 {
-				msg := "\r\n|07No files marked for download. Use |15#|07 to mark files.|07\r\n"
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte(msg)), st.outputMode)
-				time.Sleep(1 * time.Second)
-				continue // Go back to file list display
+			logoff, dlErr := st.handleFileDownload()
+			if logoff {
+				return nil, "LOGOFF", dlErr
 			}
-
-			// 2. Confirm download
-			confirmPrompt := fmt.Sprintf("Download %d marked file(s)?", len(st.currentUser.TaggedFileIDs))
-			// Use WriteProcessedBytes for SaveCursor, positioning, and clear line
-			// Need to position this prompt carefully, perhaps near the bottom prompt line.
-			// For now, just display it after the main prompt. TODO: Improve positioning.
-			terminalio.WriteProcessedBytes(st.terminal, []byte(ansi.SaveCursor()), st.outputMode)
-			terminalio.WriteProcessedBytes(st.terminal, []byte("\r\n\x1b[K"), st.outputMode) // Newline, clear line
-
-			proceed, err := st.e.PromptYesNo(st.s, st.terminal, confirmPrompt, st.outputMode, st.nodeNumber, st.termWidth, st.termHeight, false)
-			terminalio.WriteProcessedBytes(st.terminal, []byte(ansi.RestoreCursor()), st.outputMode) // Restore cursor after prompt
-
-			if err != nil {
-				if errors.Is(err, io.EOF) {
-					slog.Info("user disconnected during download confirmation", "node", st.nodeNumber)
-					return nil, "LOGOFF", io.EOF
-				}
-				slog.Error("error getting download confirmation", "node", st.nodeNumber, "error", err)
-				msg := "\r\n|01Error during confirmation.|07\r\n"
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte(msg)), st.outputMode)
-				time.Sleep(1 * time.Second)
-				continue // Back to file list
-			}
-
-			if !proceed {
-				slog.Debug("user cancelled download", "node", st.nodeNumber)
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|07Download cancelled.|07")), st.outputMode)
-				time.Sleep(500 * time.Millisecond)
-				continue // Back to file list
-			}
-
-			// 3. Protocol selection
-			proto, protoOK, protoErr := st.e.selectTransferProtocol(st.s, st.terminal, st.outputMode)
-			if protoErr != nil {
-				if errors.Is(protoErr, io.EOF) {
-					return nil, "LOGOFF", protoErr
-				}
-				slog.Error("protocol selection error", "node", st.nodeNumber, "error", protoErr)
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|01Error: No transfer protocols configured on this system.|07\r\n")), st.outputMode)
-				time.Sleep(2 * time.Second)
-				continue
-			}
-			if !protoOK {
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|07Download cancelled.|07\r\n")), st.outputMode)
-				time.Sleep(500 * time.Millisecond)
-				continue
-			}
-
-			// 4. Resolve tagged files to paths; pre-count lookup failures.
-			type dlEntry struct {
-				id   uuid.UUID
-				path string
-				name string
-			}
-			var resolved []dlEntry
-			var successCount, failCount int
-			for _, fileID := range st.currentUser.TaggedFileIDs {
-				filePath, pathErr := st.e.FileMgr.GetFilePath(fileID)
-				if pathErr != nil {
-					slog.Error("failed to get path for file", "node", st.nodeNumber, "fileID", fileID, "error", pathErr)
-					failCount++
-					continue
-				}
-				if _, statErr := os.Stat(filePath); statErr != nil {
-					slog.Error("file not on disk", "node", st.nodeNumber, "path", filePath, "fileID", fileID, "error", statErr)
-					failCount++
-					continue
-				}
-				resolved = append(resolved, dlEntry{id: fileID, path: filePath, name: filepath.Base(filePath)})
-			}
-
-			if len(resolved) == 0 {
-				slog.Warn("no valid file paths found for tagged files", "node", st.nodeNumber)
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|01Could not find any of the marked files on the server.|07\r\n")), st.outputMode)
-				failCount = len(st.currentUser.TaggedFileIDs)
-			} else {
-				paths := make([]string, len(resolved))
-				fileIDs := make([]uuid.UUID, len(resolved))
-				for i, fe := range resolved {
-					paths[i] = fe.path
-					fileIDs[i] = fe.id
-				}
-				transferSuccess, transferFail := st.e.runTransferSend(st.s, st.terminal, proto, paths, fileIDs, st.outputMode, st.nodeNumber)
-				successCount += transferSuccess
-				failCount += transferFail
-				time.Sleep(1 * time.Second)
-			}
-
-			// 4. Clear tags, update download count, and save user state
-			slog.Debug("clearing tagged file IDs", "node", st.nodeNumber, "count", len(st.currentUser.TaggedFileIDs), "handle", st.currentUser.Handle)
-			st.currentUser.TaggedFileIDs = nil // Clear the list
-			st.currentUser.NumDownloads += successCount
-			if err := st.userManager.UpdateUser(st.currentUser); err != nil {
-				slog.Error("failed to save user data after download attempt", "node", st.nodeNumber, "error", err)
-				// Inform user? State might be inconsistent.
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|01Error saving user state after download.|07")), st.outputMode)
-			}
-
-			// 5. Final status message
-			statusMsg := fmt.Sprintf("|07Download attempt finished. Success: %d, Failed: %d.|07\r\n", successCount, failCount)
-			terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte(statusMsg)), st.outputMode)
-			time.Sleep(2 * time.Second)
-
-			// Go back to the file list (will redraw with cleared marks)
 			continue
 		case "U": // Upload Files
-			slog.Debug("upload command entered", "node", st.nodeNumber, "area", st.currentAreaID, "tag", st.currentAreaTag)
-			uploadErr := st.e.runUploadFiles(st.s, st.terminal, st.currentUser, st.userManager, st.currentAreaID, st.currentAreaTag, st.outputMode, st.nodeNumber, st.sessionStartTime)
-			if uploadErr != nil {
-				if errors.Is(uploadErr, io.EOF) {
-					return nil, "LOGOFF", uploadErr
-				}
-				slog.Error("upload failed", "node", st.nodeNumber, "error", uploadErr)
+			logoff, upErr := st.handleFileUpload()
+			if logoff {
+				return nil, "LOGOFF", upErr
 			}
-			// Reload user to get updated NumUploads
-			if reloaded, exists := st.userManager.GetUser(st.currentUser.Handle); exists {
-				st.currentUser = reloaded
-			}
-			// Refresh file count and page data
-			st.totalFiles, _ = st.e.FileMgr.GetFileCountForArea(st.currentAreaID)
-			if st.filesPerPage > 0 {
-				st.totalPages = (st.totalFiles + st.filesPerPage - 1) / st.filesPerPage
-			}
-			if st.totalPages == 0 {
-				st.totalPages = 1
-			}
-			if st.currentPage > st.totalPages {
-				st.currentPage = st.totalPages
-			}
-			st.filesOnPage, _ = st.e.FileMgr.GetFilesForAreaPaginated(st.currentAreaID, st.currentPage, st.filesPerPage)
 			continue
 		case "V": // View file
-			slog.Debug("view command entered in file list", "node", st.nodeNumber)
-			viewPrompt := "\r\n|07Enter file # to view (or |15ENTER|07 to cancel): |15"
-			terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte(viewPrompt)), st.outputMode)
-			viewInput, viewErr := readLineFromSessionIH(st.s, st.terminal)
-			if viewErr != nil {
-				if errors.Is(viewErr, io.EOF) {
-					return nil, "LOGOFF", io.EOF
-				}
-				continue
-			}
-			viewNum := strings.TrimSpace(viewInput)
-			if viewNum == "" {
-				continue
-			}
-			fileNumToView, parseErr := strconv.Atoi(viewNum)
-			if parseErr != nil || fileNumToView <= 0 {
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|01Invalid file number.|07\r\n")), st.outputMode)
-				time.Sleep(500 * time.Millisecond)
-				continue
-			}
-			viewIndex := fileNumToView - 1 - (st.currentPage-1)*st.filesPerPage
-			if viewIndex < 0 || viewIndex >= len(st.filesOnPage) {
-				terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|01File number not on current page.|07\r\n")), st.outputMode)
-				time.Sleep(500 * time.Millisecond)
-				continue
-			}
-			fileToView := st.filesOnPage[viewIndex]
-			if st.e.FileMgr.IsSupportedArchive(fileToView.Filename) {
-				viewFilePath, pathErr := st.e.FileMgr.GetFilePath(fileToView.ID)
-				if pathErr != nil {
-					slog.Error("failed to get path for file", "node", st.nodeNumber, "fileID", fileToView.ID, "error", pathErr)
-					terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte("\r\n|01Error locating file.|07\r\n")), st.outputMode)
-					time.Sleep(1 * time.Second)
-				} else {
-					ctx, cancel := st.e.transferContext(st.s.Context())
-					ziplab.RunZipLabView(ctx, st.s, st.terminal, viewFilePath, fileToView.Filename, st.outputMode, sessionReadLine(st.s, st.terminal), sessionReadKey(st.s))
-					cancel()
-				}
-			} else {
-				viewFileByRecord(st.e, st.s, st.terminal, &fileToView, st.outputMode, st.termWidth, st.termHeight)
+			logoff, vErr := st.handleFileView()
+			if logoff {
+				return nil, "LOGOFF", vErr
 			}
 			continue
 		case "A": // Area Change (Placeholder/Not implemented here, handled by menu?)
@@ -521,44 +344,7 @@ func runListFiles(c *cmdCtx, args string) (*user.User, string, error) {
 			terminalio.WriteProcessedBytes(st.terminal, ansi.ReplacePipeCodes([]byte(msg)), st.outputMode)
 			time.Sleep(1 * time.Second)
 		default: // Includes 'T' (Tagging) and potential numeric input
-			// Try to parse as a number for tagging
-			fileNumToTag, err := strconv.Atoi(upperInput)
-			if err == nil && fileNumToTag > 0 {
-				// Valid number entered, attempt to tag/untag
-				fileIndex := fileNumToTag - 1 - (st.currentPage-1)*st.filesPerPage
-				if fileIndex >= 0 && fileIndex < len(st.filesOnPage) {
-					fileToToggle := st.filesOnPage[fileIndex]
-					found := false
-					newTaggedIDs := []uuid.UUID{}
-					if st.currentUser.TaggedFileIDs != nil {
-						for _, taggedID := range st.currentUser.TaggedFileIDs {
-							if taggedID == fileToToggle.ID {
-								found = true // Mark as found to skip adding it back
-							} else {
-								newTaggedIDs = append(newTaggedIDs, taggedID)
-							}
-						}
-					}
-					if !found {
-						// File was not tagged, so add it
-						newTaggedIDs = append(newTaggedIDs, fileToToggle.ID)
-						slog.Debug("user tagged file", "node", st.nodeNumber, "handle", st.currentUser.Handle, "fileNum", fileNumToTag, "fileID", fileToToggle.ID)
-					} else {
-						// File was tagged, so we removed it (untagged)
-						slog.Debug("user untagged file", "node", st.nodeNumber, "handle", st.currentUser.Handle, "file_num", fileNumToTag, "id", fileToToggle.ID)
-					}
-					st.currentUser.TaggedFileIDs = newTaggedIDs
-					// No page change needed, loop will redraw with updated marks
-				} else {
-					// Invalid file number for current page
-					slog.Debug("invalid file number entered", "node", st.nodeNumber, "file_num", fileNumToTag)
-					// Optional: Add user feedback message
-				}
-			} else {
-				// Input was not N, P, Q, D, U, V, A, or a valid number - Invalid command
-				slog.Debug("invalid command entered in LISTFILES", "node", st.nodeNumber, "input", upperInput)
-				// Optional: Add user feedback message
-			}
+			st.toggleFileTag(upperInput)
 		} // end switch
 	} // end for loop
